Add a canonical url into the page meta
In order to provide a proper url property for open graph, i.e.
Facebook, we must use the original request as the source. This makes
it easy to apply this meta information on the server side.

In order to complete TH-178, we only need this meta information on the
events page, but with this commit we are applying it to all responses.

I've also applied more generic canonical url link, which is used by
search engines.