Change element focus when user navigates with keyboard
The item the user chooses receives focus. This means that the user can
select the item based by the convetions of their platform. For
browsers, the space key will select an item.

Enter no longer selects an item.

As the focus is moved from the input to the selected element, the user
is no longer able to alter they search without refocusing the input
element.